feat(proto): regenerate against upstream 0.1.9 (WP-1)
The schema was hand-transcribed from upstream 0.1.1 and had drifted eight
releases. scripts/gen_proto.py now decodes the serialized FileDescriptorProto
out of any wheel's localharness_pb2.py and renders it, so the schema is
generated rather than retyped -- which is how UsageMetadata's counters ended up
as int32 where the harness has always declared uint64.

Two deliberate divergences, documented in the script: rendered as proto3 rather
than editions, because prost-build 0.12 cannot parse editions; and every
singular field marked optional, because the step classifier and the request
dedup are built on .is_some() and losing presence would turn "absent" into
"present and zero". content.proto is not vendored -- the only fields
referencing it are ToolCall.arguments and ToolResponse.response, which neither
SDK populates.

build.rs sets pbjson's ignore_unknown_fields, so a newer harness adding a field
no longer discards the entire event. It does NOT cover unknown enum variants,
which is why the idle rename below had to come from regenerating.

The regeneration produced 16 compile errors -- the intended forcing function.
Resolved here:

  - HarnessConfig.gemini_config (field 2) is gone; src/harness_config.rs now
    builds the repeated ModelConfig models (field 15) that replaced it in
    0.1.4, emitting the two-entry TEXT+IMAGE shape upstream produces. The
    public GeminiConfig type is unchanged: this gets the wire right without an
    API break, leaving the ModelTarget rewrite to WP-4.
  - UsageMetadata counters int32 -> u64, matching the proto.
  - Added ClientInfo.os/os_version, InputConfig.env, Tool.defer_loading, the
    three HarnessSideTools configs, ToolResponse.error_message and
    ActionGenerateImage.aspect_ratio. Fields belonging to later work packages
    are explicitly unset with a comment naming the package, so the compiler
    flags them again when those land.
  - Removed GenerateImageToolConfig.model_name, reserved upstream in 0.1.4.
  - Handled the three new OutputEvent arms. call_hook_request logs a warning
    that the harness may stall, because arriving there means enabled_hooks and
    the router have gone out of sync and we cannot yet answer it.

Both mocks spoke the pre-0.1.9 dialect and the suite caught it exactly as the
audit predicted a real harness would behave: STATE_IDLE was renamed
STATE_FULLY_IDLE, protojson matches on the value name, so the idle event was
dropped as an unknown variant and test_wasm_connection_integration_mock hung
forever instead of ending its turn. STATE_TERMINAL_ERROR = 5 was likewise
removed upstream in 0.1.3; a failing step now reports STATE_ERROR.

The SDK still does not read the initialize_conversation_response handshake
frame (WP-6), so it cannot yet complete a connection to a real 0.1.9 harness.

Verified with the CI toolchain: clippy --all-targets --all-features -D
warnings, fmt --check, and 129 tests pass.

+++ src/types.rs
@@ -437,17 +437,17 @@ pub struct ToolResult {
 #[derive(Debug, Clone, Serialize, Deserialize, Default)]
 pub struct UsageMetadata {
     /// Tokens included in the request prompt.
-    pub prompt_token_count: i32,
+    pub prompt_token_count: u64,
     /// Tokens generated in candidates.
-    pub candidates_token_count: i32,
+    pub candidates_token_count: u64,
     /// Total combined tokens.
-    pub total_token_count: i32,
+    pub total_token_count: u64,
     /// Cache hit content tokens.
     #[serde(default)]
-    pub cached_content_token_count: i32,
+    pub cached_content_token_count: u64,
     /// Tokens consumed during inner thinking/reasoning.
     #[serde(default)]
-    pub thoughts_token_count: i32,
+    pub thoughts_token_count: u64,
 }
 
 /// The classification type of a step in the trajectory.
